Unify ScMidiMessage tests into a single test class per convention
